Remove SamlEntityRepositoryInterface postfix
